Merging comment updates with "normal" work item updates.
When a comment is updated, we actually get a work item update, too. But it
contains an unusable "System.History" change. Still, we have a work item
update. And, if there are other fields updated simultaneously, we also
printed an update table for it. Having two update tables, one for the
work item update and one for the comment update, at the same date is
unnecessary. So merge them.

Also fixed a few typescript check errors.