Improved detection of light/dark mode.
Found that the undocumented event 'themeApplied' exists. It gets called
after the theme is applied. So detecting the theme in it makes the most
sense. Moreover, it also gets called when the user changes the theme,
meaning that now we can adapt our colors immediately, too. (Before, when
switching from e.g. light to dark mode, the ins/del colors were wrong
until a page reload.)