Add a vss-extension.json version for private testing in the cloud-based ADO Services.
To test the extension in the Microsoft hosted Azure DevOps Services, one needs to upload and publish the extension to the marketplace. To do this privately, one needs to upload the extension under a different name and with the "public" flag set to "false". This is done by the manifest committed here.
Differences compared to the main manifest vss-extension.json:
- Different extension name
- public=false
- History tab name contains a "Test"