Show the work item's revision and stack rank in the history.
Stack rank: The stack rank is a large integer that specifies the order of items in boards/backlogs. The absolute value is without meaning, but its meaning is in relation to other work items. Hence I thought it is never interesting to users. But as explained in GitHub issue #2, an increase or decrease of the value itself also holds some meaning (i.e. whether a work item was made more or less important). Also, it might be confusing for users if some work item update numbers get skipped. Given that since we nowadays allow users to configure filters, they can filter the stack rank out themselves.

Revision field "Rev": Use case from GitHub issue #3: If a specific version of a work item gets stored/printed/... externally together with the work item's revision (e.g. as some form of requirement document), it would be nice to see the revision in the history so that users can match them. ADO itself only allows showing the revision in queries, but not directly on a work item. As for the stack rank, users can filter out the "Rev" field themselves nowadays. Also, I thought about hardcoding the friendly name as "Revision", but decided against it considering that ADO itself uses "Rev" in queries.

Compare GitHub issues #2 and #3.