Reduce flicker during loading of the config buttons.
Especially when switching between different work items in ADO services (cloud), it takes like half a second for the config to load. So, the buttons get rendered first with the text from historydiff.html. And half a second later they are set properly. This is distracting. So show the buttons only once we know the actual values to display.
It also makes more sense in case of very long loading times, since we don't want to let the user click the buttons before the config has been loaded.