Position the two config buttons next to the topmost tile if there is enough space.
I.e. if the max. width of the tiles is limited and there is sufficient space between the right border of the tile and the window tile, then place the buttons there.
Otherwise, keep the current behavior where the buttons are placed at the top right corner of the document and are allowed to overlap with the tiles.