Fixed config buttons sometimes incorrectly overlapping with the update tile.
Sometimes, the avatar image in the heading of the update tile loads delayed, i.e. only after we called RepositionConfigButtons(). In that case, the tile will change width again, causing the buttons to be positioned incorrectly.
Solution: Ensure that the browser reserves the correct image width from the very beginning. Since avatars are square, it doesn't matter for the image itself whether we set the width or height.