Implemented a "please rate" notice.
It is shown 30 days after the user first used the extension.
I put so much time and effort into the extension, and it is free, so at least politely request a positive review and a GitHub star as a reward for me :-)
More stars/reviews should also help to boost its popularity.