Try to detect if the old or new value of a markdown field actually contains html.
The info that a field is in markdown is for the latest revision of the work item. It might have been in html in the past. So attempt to detect html content in past revisions for proper html diffs.
This isn't foolproof, as explained in the comment, but hopefully good enough for most cases.

See issue #4.